temporary nav along with router for pages
